+++ src/main/resources/templates/index.html
@@ -1,12 +1,20 @@
 <!DOCTYPE html>
 <html lang="en" xmlns:th="http://www.thymeleaf.org/">
-<head>
-    <meta charset="UTF-8"/>
-    <title></title>
-</head>
-<body>
+<head th:replace="fragments :: head"></head>
 
-<p>Hello World!</p>
-
+<body class ="container">
+<h1>Things to do... </h1>
+<nav th:replace="fragments :: navigation"></nav>
+<p th:unless ="${items} and ${items.size()}">Nothing to do.</p>
+<table class="table table-striped">
+    <thead>
+    <tr>
+        <th>Name</th>
+    </tr>
+    </thead>
+    <tr th:each="items : ${items}">
+        <td th:text="${items}"></td>
+    </tr>
+</table>
 </body>
-</html>
+</html>
